refactor: fix imports after type reorganization
Routes server imports (AppConfig, GeminiInlineData, GeminiRequest,
GeminiFunctionDeclaration, DriveFileInfo) from shared/types to
server/types, and client RecipeParams from shared/types to client/types.
Adds temporary casts and SavedState adjustment to keep typecheck clean
until Tasks 5/7/8 refactor the dependent logic.
