+++ src/server/tools.ts
@@ -1,4 +1,4 @@
 // src/server/tools.ts
-import type { GeminiFunctionDeclaration } from "../shared/types";
+import type { GeminiFunctionDeclaration } from "./types";
 
 export const TOOL_REGISTRY: Record<string, GeminiFunctionDeclaration> = {};
